feat(storage): read data from sstable
And create general index

+++ packages/chest/src/tests.rs
@@ -9,7 +9,9 @@ fn ensure_dir_exists(dir_path: &PathBuf) -> std::io::Result<()> {
 
 fn get_test_tempdir(custom_name: &str) -> PathBuf {
     let path = std::env::temp_dir().join(custom_name);
-    std::fs::remove_dir_all(&path).unwrap();
+    if path.exists() {
+        std::fs::remove_dir_all(&path).unwrap();
+    }
     ensure_dir_exists(&path).unwrap();
     path
 }
@@ -33,4 +35,14 @@ fn test_flush() {
     assert_eq!(chest.len(), 1);
     chest.set("age", Value::Integer(5));
     assert_eq!(chest.len(), 0);
+}
+#[test]
+fn test_read_from_sstable() {
+    let chest_dir = get_test_tempdir("dungeon-testing-3");
+    let mut chest = Chest::new(chest_dir, 2);
+    chest.set("foo", Value::String("bar".to_string()));
+    chest.set("foo2", Value::String("bar2".to_string()));
+    assert_eq!(chest.len(), 0);
+    assert_eq!(chest.get("foo"), Some(Value::String("bar".to_string())));
+    assert_eq!(chest.get("foo2"), Some(Value::String("bar2".to_string())));
 }
